specify printer model with model attribut
previously the printer model was specified with the `driver` attribute.
as there is a `driver` attribute in the `cups_lpadmin` module which does
not specifie the model of the printer the the `model` attribute was
added to specifie the model and the `driver` attribute repurposed to do
the same as in the `cups_lpadmin` module.